Fixed iteration loop instead of enumeration
Although the for..in is usually correct, it's actually an enumeration instead of a true iteration.
The problem came about when using prototype and jQuery.noConflict().

Prototype by default adds a few extra methods to the Array, enumerating through the array, shows the array method, which the plugin tries to render as bars.
This causes a lot of 'undefined' values to show up. Doing a proper iteration through the jQuery $.each() method makes it work as it should.